+++ PI_BalderramaEnrique/BackEnd/src/main/java/com/portfolio/lebb/Service/ImpPersonaService.java
@@ -30,12 +30,12 @@ public class ImpPersonaService{
     }
 
     
-    public void deletePersona(Long id) {
+    public void deletePersona(int id) {
        ipersonaRepository.deleteById(id);
     }
 
     
-    public Persona findPersona(Long id) {
+    public Persona findPersona(int id) {
         Persona persona = ipersonaRepository.findById(id).orElse(null);
         return persona;
     }
